search unit tests done
