+++ babel.config.js
@@ -15,6 +15,7 @@ module.exports = {
         plugins: [
          '@babel/plugin-transform-react-jsx-source',
         '@babel/plugin-transform-react-jsx-self',
+           '@babel/plugin-transform-runtime',
         ]
       }
   }
