+++ src/api/bestPosts/BestPosts.module.css
@@ -1,0 +1,9 @@
+.div{
+    position: relative;
+    top: 10vh;
+}
+
+.ul{
+    list-style: none;
+    padding: 0;
+}
