+++ src/api/popularPosts/PopularPosts.module.css
@@ -1,0 +1,14 @@
+.div{
+    position: relative;
+    top: 10vh;
+}
+
+.ul{
+    list-style: none;
+    padding: 0;
+}
+
+.li{
+    list-style: none;
+
+}
