+++ src/store/loadingSlice.js
@@ -2,7 +2,6 @@ import { createSlice } from "@reduxjs/toolkit";
 
 const initialState = {
     loading: false,
-    completed: false,
     error:false
 
 }
@@ -13,17 +12,14 @@ const loadingSlice = createSlice({
     reducers: {
         changeLoadingState(state,action){
             state.loading = true
-            state.completed = false
             state.error = false
         },
         changeCompletedState(state,action){
             state.loading = false
-            state.completed = true
             state.error = false
         },
         changeErrorState(state,action){
             state.loading = false
-            state.completed = false
             state.error = true
         },
     }
