+++ src/tests/popularPostsTests/unitTests/popularPosts/getPopularPosts.test.js
@@ -2,7 +2,7 @@ import Enzyme from 'enzyme';
 import { EnzymeAdapter } from 'enzyme';
 //import 'cross-fetch/polyfill'
 import * as popularPostsData from './PopularPostsData'
-import * as sortPosts from '../../../../helpers/SortReceivedPosts'
+import * as sortPosts from '../../../../helpers/sortReceivedPosts/SortReceivedPosts'
 import * as popular from '../../../../api/popularPosts/getPopularPosts'
 import * as nestedObjectFieldsValidation from '../../../../helpers/nestedObjectsFieldValidation/getNestedObjects';
 import 'core-js/stable/structured-clone';
