+++ src/tests/search/unitTests/searchPostsData.js
@@ -1,0 +1,127 @@
+export const searchData = {
+    oneValidSearchPosts: {
+        "kind": "Listing",
+        "data": {
+            "modhash": "",
+            "dist": 25,
+            "facets": {},
+            "after": "t3_155olnd",
+            "geo_filter": "",
+            "children": [
+                {
+                    "kind": "t3",
+                    "data": {
+                        subreddit_name_prefixed: "1",
+                        selftext: "1",
+                        title: "1",
+                        score: "1",
+                        name: "1",
+                        ups: "1",
+                        downs: "1",
+                        imageUrl: "1",
+                        post_hint: "1",
+                        subreddit_subscribers: "1",
+                        id: "1",
+                        num_comments: "1",
+                        author: "1",
+                    }
+                }
+            ]
+        }
+    },
+    twoValidSearchPosts: {
+        "kind": "Listing",
+        "data": {
+            "modhash": "",
+            "dist": 25,
+            "facets": {},
+            "after": "t3_155olnd",
+            "geo_filter": "",
+            "children": [
+                {
+                    "kind": "t3",
+                    "data": {
+                        subreddit_name_prefixed: "1",
+                        selftext: "1",
+                        title: "1",
+                        score: "1",
+                        name: "1",
+                        ups: "1",
+                        downs: "1",
+                        imageUrl: "1",
+                        post_hint: "1",
+                        subreddit_subscribers: "1",
+                        id: "1",
+                        num_comments: "1",
+                        author: "1",
+                    }
+                },
+                {
+                    "kind": "t3",
+                    "data": {
+                        subreddit_name_prefixed: "2",
+                        selftext: "2",
+                        title: "2",
+                        score: "2",
+                        name: "2",
+                        ups: "2",
+                        downs: "2",
+                        imageUrl: "2",
+                        post_hint: "2",
+                        subreddit_subscribers: "2",
+                        id: "2",
+                        num_comments: "2",
+                        author: "2",
+                    }
+                }
+            ]
+        }
+    },
+    oneSortedPost: [{
+            subreddit_name_prefixed: "1",
+            selftext: "1",
+            title: "1",
+            score: "1",
+            name: "1",
+            ups: "1",
+            downs: "1",
+            imageUrl: "1",
+            post_hint: "1",
+            subreddit_subscribers: "1",
+            id: "1",
+            num_comments: "1",
+            author: "1",
+        }],
+    twoSortedPosts: [{
+        subreddit_name_prefixed: "1",
+        selftext: "1",
+        title: "1",
+        score: "1",
+        name: "1",
+        ups: "1",
+        downs: "1",
+        imageUrl: "1",
+        post_hint: "1",
+        subreddit_subscribers: "1",
+        id: "1",
+        num_comments: "1",
+        author: "1",
+    },
+        {
+            subreddit_name_prefixed: "2",
+            selftext: "2",
+            title: "2",
+            score: "2",
+            name: "2",
+            ups: "2",
+            downs: "2",
+            imageUrl: "2",
+            post_hint: "2",
+            subreddit_subscribers: "2",
+            id: "2",
+            num_comments: "2",
+            author: "2",
+        }
+    ]
+
+}
