--- src/tests/sortReceivedPostAndCommentsTests/SortReceivedPostAndCommentsDataExpectedResults.js
@@ -1,16 +1,0 @@
-export const validCommentDataExpectedResults = {
-    commentWithCorrectFields: {
-        isValid: true,
-        currentComment: {
-            id: "w",
-            author: "author",
-            createdUTC: "created utc date",
-            downs: "downs",
-            ups: "ups",
-            body: "commentWithCorrectFields",
-            subreddit_name_prefixed: "subreddit_name_prefixed",
-            replies: [],
-            moreReplies: []
-        }}
-
-}
